requirements.txt
llama-index
python-dotenv
llama-index-llms-google-genai
llama-index-embeddings-google-genai
llama-index-vector-stores-weaviate
pypdf
weaviate-client
Flask
Flask-Cors
scrapy
requests
beautifulsoup4
lxml
python-docx
openpyxl
python-pptx
PyMuPDF
transformers
torch
watchfiles>=0.21.0
easyocr